Finish landmark search control
