--- BnbnavNetClient/Views/LandmarkSearchView.axaml
@@ -1,30 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="BnbnavNetClient.Views.LandmarkSearchView"
-             x:DataType="viewModels:LandmarkSearchControlViewModel"
-             xmlns:str="using:BnbnavNetClient.I18Next"
-             xmlns:viewModels="clr-namespace:BnbnavNetClient.ViewModels"
-             x:CompileBindings="True">
-    <StackPanel>
-        <TextBox Watermark="{str:Tr Key=SEARCH}" Text="{Binding SearchQuery}" VerticalAlignment="Top">
-            <TextBox.InnerRightContent>
-                <Button Content="X" Command="{Binding ClearSearchQuery}"></Button>
-            </TextBox.InnerRightContent>
-        </TextBox>
-        <StackPanel IsVisible="{Binding SelectedLandmark, Converter={x:Static ObjectConverters.IsNull}}" VerticalAlignment="Stretch">
-            <ListBox Items="{Binding SearchResults}" SelectedItem="{Binding SelectedLandmark}">
-                <ListBox.ItemTemplate>
-                    <DataTemplate>
-                        <StackPanel>
-                            <TextBlock Text="{Binding Name}"></TextBlock>
-                            <TextBlock Text="{Binding HumanReadableType}"></TextBlock>
-                        </StackPanel>
-                    </DataTemplate>
-                </ListBox.ItemTemplate>
-            </ListBox>
-        </StackPanel>
-    </StackPanel>
-</UserControl>
